+++ src/Command/BuildHooksCommand.php
@@ -37,7 +37,7 @@ final class BuildHooksCommand extends Command
 		// @todo remove code block + alias in v4
 		if ("hosting:hook:build" !== $input->getFirstArgument())
 		{
-			$message = sprintf(
+			$message = \sprintf(
 				"The command `%s` was deprecated. Use `hosting:hook:build` instead.",
 				$input->getFirstArgument(),
 			);