+++ src/Deployment/TaskCli.php
@@ -11,7 +11,7 @@ final class TaskCli extends TorrStyle
 	 */
 	public function done (string $message) : void
 	{
-		$this->write(sprintf(
+		$this->write(\sprintf(
 			"<fg=green>✓</> %s",
 			$message,
 		));